% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_scores.R
\name{combine_scores}
\alias{combine_scores}
\title{Combines standardized scores}
\usage{
combine_scores(df_rna = NULL, df_prot = NULL, df_kin = NULL, tf_kin = NULL)
}
\arguments{
\item{df_rna}{dataframe with percentile ranking of RNA features}

\item{df_prot}{dataframe with percentile ranking of protein features}

\item{df_kin}{dataframe with percentile ranking of kinase features}

\item{tf_kin}{dataframe with percentile ranking of transcription factor features}
}
\value{
dataframe with combined scores
}
\description{
This function ranks combines standardized scores from each omic ranked dataset
}
\examples{
TRUE
}
