% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinograte.R
\name{kinograte}
\alias{kinograte}
\title{Integrate omic datasets using PCSF}
\usage{
kinograte(
  df,
  ppi_network,
  n = 10,
  w = 10,
  r = 0.1,
  b = 2,
  mu = 0.005,
  cluster = TRUE,
  seed = NULL
)
}
\arguments{
\item{df}{combined ranked omic datasets}

\item{ppi_network}{dataframe of protein-protein interactions}

\item{n}{An integer value to determine the number of runs with random noise added edge costs. A default value is 10.}

\item{w}{A numeric value for tuning the number of trees in the output. A default value is 2.}

\item{r}{A numeric value to determine additional random noise to edge costs. A random noise upto r percent of the edge cost is added to each edge. A default value is 0.1}

\item{b}{A numeric value for tuning the node prizes. A default value is 1.}

\item{mu}{A numeric value for a hub penalization. A default value is 0.0005.}

\item{cluster}{set TRUE to cluster the network}

\item{seed}{(optional) set seed number}
}
\value{
list(network, nodes, edges)
}
\description{
This function integrates omic datasets using prize-collecting Steiner forest (PCSF) algorithm
}
