% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kinomics_exmaple}
\alias{kinomics_exmaple}
\title{an example of Kinome data}
\format{
A data frame with 234 rows and 2 variables:
\describe{
\item{hgnc_symbol}{Gene Symbols of kinases}
\item{score}{normlaized score}
}
}
\usage{
kinomics_exmaple
}
\description{
an example of Kinome data
}
\keyword{datasets}
