% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_enrichment.R
\name{network_enrichment}
\alias{network_enrichment}
\title{Network pathway analysis}
\usage{
network_enrichment(network, ...)
}
\arguments{
\item{network}{integrated network}

\item{...}{arguments passed to PCSF::enrichment_analysis()}
}
\value{
df enriched terms
}
\description{
This function performs pathway enrichment analysis on the intergated network
}
