% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentile_rank.R
\name{percentile_rank}
\alias{percentile_rank}
\title{Ranks genes/proteins using a percentile ranking for a selected variable}
\usage{
percentile_rank(df, symbol, metric, desc = FALSE)
}
\arguments{
\item{df}{dataframe that contains genes/proteins to rank (tidy format)}

\item{symbol}{column name that contains the gene/protein symbols}

\item{metric}{column name to be used as the metric to rank}

\item{desc}{boolean, ranking in a descending or ascending order. Default = FALSE}
}
\value{
dataframe with percentile ranking
}
\description{
This function ranks genes or proteins using a percentile ranking for a selected variable (for example, fold change or pvalue)
}
\examples{
TRUE
}
