% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppi_network_example}
\alias{ppi_network_example}
\title{an example of ppi network}
\format{
A data frame with 175205 rows and 3 variables:
\describe{
\item{head}{Protein 1}
\item{tail}{Protein 2}
\item{cost}{inverse degree of confidence}
}
}
\usage{
ppi_network_example
}
\description{
an example of ppi network
}
\keyword{datasets}
