% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proteomics_exmaple}
\alias{proteomics_exmaple}
\title{an example of proteomics data}
\format{
A data frame with 141 rows and 2 variables:
\describe{
\item{Gene.Symbol}{Gene Symbols}
\item{LFC}{log2 fold change}
}
}
\usage{
proteomics_exmaple
}
\description{
an example of proteomics data
}
\keyword{datasets}
