% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rnaseq_example}
\alias{rnaseq_example}
\title{an example of RNA differential gene expression data}
\format{
A data frame with 3207 rows and 3 variables:
\describe{
\item{Gene name}{Gene Symbols}
\item{LFC}{log2 fold change}
\item{P_Value}{pvalue}
}
}
\usage{
rnaseq_example
}
\description{
an example of RNA differential gene expression data
}
\keyword{datasets}
