% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_hits.R
\name{top_hits}
\alias{top_hits}
\title{Extracts top hits (genes/proteins) based on the normalized score}
\usage{
top_hits(df, prec_cutoff, omic_type)
}
\arguments{
\item{df}{dataframe that contains ranked genes/proteins}

\item{prec_cutoff}{the percentile cutoff}

\item{omic_type}{name of omic dataset (eg. RNA, Protein, Kinase, ... etc)}
}
\value{
dataframe of the top hits
}
\description{
This function extracts the top hits (genes/proteins) based on the normalized score
which is the percentile rank using an adjustable cutoff
}
\examples{
TRUE

}
