% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_network.R
\name{visualize_network}
\alias{visualize_network}
\title{Visualize integrated network}
\usage{
visualize_network(
  nodes,
  edges,
  cluster_df = NULL,
  layout = "layout_with_fr",
  seed = 123,
  options_by = "group"
)
}
\arguments{
\item{nodes}{network nodes}

\item{edges}{network edges}

\item{layout}{layout option from igraph. default = "layout_with_fr". \href{https://igraph.org/r/doc/layout_.html}{See full list}}

\item{seed}{(optional) set seed}

\item{options_by}{(optional) set dropdown menu, "group" or "cluster". Set NULL to remove dropdown menu}
}
\value{
visNetwork object
}
\description{
This function visualizes the integrated results with an interactive network
}
